First pass ready to test

+++ src/com/example/simplecolab/SimpleComputeTaskListener.java
@@ -5,7 +5,7 @@ public interface SimpleComputeTaskListener {
 	 * This class is a  listener for events from the compute asynch task
 	 */
 	
-	public void onSimpleComputeFinished(int result);
+	public void onSimpleComputeFinished(double result);
 	// called when the compression task has completed
 	
 	public void onSimpleCommputePorgressUpdate();
